Nested objects Normalization/Denormalization
By default, nested objects are presented by URI

{
    'id': 123,
    'username': 'jonsnow_1',
    'isActive': true,
    'roles': [
        '/roles/10'
    ]
}
But you can achieve something like this

{
    'id': 123,
    'username': 'jonsnow_1',
    'isActive': true,
    'roles': [
        {
            'symbol': 'ROLE_SUPER_ADMIN',
            'description': 'You can do whatever you want'
        }
    ]
}

This approach allows you to improve performance by avoiding execution of many queries to get whole data. In this example we can save one request but what in such a  case when we have a list of hundreds of items which have relation to hundreds of nested objects? What in such a case when we want to get everything in one query?