We can use this approach in denormalization either, when we want to add a nested object to a new or existing object. The only thing which has to be done is adding denormalization group  'account.write' and 'account.update' to properties of the nested object.
This configuration allows to:

– Add Role to existing Account

Content-Type: application/json

Method: PUT

{
   'id':123,
   'username':'jonsnow_1',
   'isActive':true,
   'roles':[
      {
         'id':10
      }
   ]
}

– Save a new Account with already created Role(s)

Content-Type: application/json

Method: POST

{
   'username':'jonsnow_new',
   'isActive':true,
   'password':'test',
   'roles':[
      {
         'id':10
      }
   ]
}

– Create a completely new Role object if there is no id in the request

Content-Type: application/json

Method: PUT

{
   'id':123,
   'username':'jonsnow_1',
   'isActive':true,
   'roles':[
      {
         'symbol':'ROLE_SUPER_ADMIN_EXTRA',
         'description':'Lorem ipsum'
      }
   ]
}

The last example is not a super safe one in our case but it shows how nested denormalization works.

If you provide Role id, then object is loaded from database and it is set as a relation to an actual Account object.

In case that you do not provide Role id, a new object along with provided parameters, is created and set as a relation to an actual Account object.